CRUD article complet

+++ backend/models/articleModel.js
@@ -29,8 +29,9 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
       defaultValue: "default/default.jpg",
     },
+    //Changer en réel, sinon pas de virgule
     prixUnitaire: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
       field: "prix_unitaire",
       allowNull: false,
     },
